requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.3
django-ace==1.0.5
django-allauth==0.39.1
django-crispy-forms==1.7.2
idna==2.8
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
oauthlib==3.0.2
Pillow==6.1.0
pkg-resources==0.0.0
protobuf==3.9.0
pydjax==0.2
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
